+++ packages/koot/ReactApp/server/middlewares/isomorphic/index.js
@@ -251,7 +251,7 @@ const middlewareIsomorphic = (options = {}) => {
 
             if (result.error) {
                 // renderComplete();
-                throw result.error;
+                throw transformError(result.error);
             }
 
             if (result.redirect) {
@@ -292,4 +292,31 @@ const extendCacheObject = (cache, chunkmap, localeId) => {
             ? devRequestServiceWorker
             : serviceWorker;
     }
+};
+
+/**
+ * 转换 Error 对象
+ * @param {Error|string} error
+ * @returns {Error} 确定会有 `message` 和 `msg` 属性的 Error 对象
+ */
+export const transformError = (error) => {
+    let msg =
+        error?.message ??
+        error?.msg ??
+        (typeof error === 'string' ? error : '网络异常，请稍后重试');
+    if (msg === 'Network Error') msg = '网络异常，请稍后重试';
+
+    if (error instanceof Error) {
+        // error = err;
+    } else if (typeof error === 'object') {
+        error = new Error(msg);
+        for (const [key, value] of Object.entries(error)) error[key] = value;
+    } else {
+        error = new Error(error);
+    }
+
+    error.message = msg;
+    error.msg = msg;
+
+    return error;
 };