+++ packages/koot/defaults/service-worker.js
@@ -4,6 +4,7 @@
 module.exports = {
     auto: true,
     filename: 'service-worker.js',
+    // scope: '/', // 生成 webpack 配置时，根据是否是 SPA 动态设置该默认值
     include: [],
-    exclude: []
+    exclude: [],
 };